+++ service/service.go
@@ -151,7 +151,6 @@ func New(config Config) (*Service, error) {
 				Provider:                provider,
 				Region:                  config.Viper.GetString(config.Flag.Service.Installation.Region),
 				Registry:                config.Viper.GetString(config.Flag.Service.Installation.Registry),
-				IngressAPIVersion:       config.Viper.GetString(config.Flag.Service.Installation.IngressAPIVersion),
 
 				OpsgenieKey: config.Viper.GetString(config.Flag.Service.Opsgenie.Key),
 
@@ -191,7 +190,6 @@ func New(config Config) (*Service, error) {
 				Provider:                provider,
 				Region:                  config.Viper.GetString(config.Flag.Service.Installation.Region),
 				Registry:                config.Viper.GetString(config.Flag.Service.Installation.Registry),
-				IngressAPIVersion:       config.Viper.GetString(config.Flag.Service.Installation.IngressAPIVersion),
 
 				OpsgenieKey: config.Viper.GetString(config.Flag.Service.Opsgenie.Key),
 
@@ -231,7 +229,6 @@ func New(config Config) (*Service, error) {
 			Provider:                provider,
 			Region:                  config.Viper.GetString(config.Flag.Service.Installation.Region),
 			Registry:                config.Viper.GetString(config.Flag.Service.Installation.Registry),
-			IngressAPIVersion:       config.Viper.GetString(config.Flag.Service.Installation.IngressAPIVersion),
 
 			AlertmanagerAddress:     config.Viper.GetString(config.Flag.Service.Alertmanager.Address),
 			AlertmanagerBaseDomain:  config.Viper.GetString(config.Flag.Service.Alertmanager.BaseDomain),