+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve HasPrometheusAgent function to ignore prometheus-agent scraping targets.
 - Do not reconcile service monitors in CAPO MCs.
 
+### Removed
+
+- Remove deprecated ingress v1beta1 (only used on kvm).
+
 ## [4.12.0] - 2022-11-25
 
 ### Changed